+++ packages/manager/src/features/Firewalls/FirewallDetail/Devices/FirewallDeviceLanding.tsx
@@ -6,8 +6,7 @@ import * as React from 'react';
 
 import { DebouncedSearchTextField } from 'src/components/DebouncedSearchTextField';
 
-//import { AddLinodeDrawer } from './AddLinodeDrawer';
-import { AddLinodeDrawer } from './AddLinodeDrawerNoUseEffect';
+import { AddLinodeDrawer } from './AddLinodeDrawer';
 import { AddNodebalancerDrawer } from './AddNodebalancerDrawer';
 import { formattedTypes } from './constants';
 import { FirewallDeviceTable } from './FirewallDeviceTable';
